Add Editor to index page. Fixed bugs and minor visual adjustment
