+++ src/components/Index/index.js
@@ -1,10 +1,11 @@
 import React, { Component } from 'react';
-
+import Editor from '../editor/Editor'
 class Index extends Component {
   render() {
     return (
       <div>
         <h1>Index Page</h1>
+        <Editor />
       </div>
     );
   }
