+++ src/components/editor/BlockStyleControls.js
@@ -3,12 +3,12 @@ import StyleButton from './StyleButton'
 
 
 const BLOCK_TYPES = [
-  {label: 'H1', style: 'header-one', icon:'fa fa-header'},
-  {label: 'H2', style: 'header-two', icon:'fa fa-header'},
-  {label: 'H3', style: 'header-three', icon:'fa fa-header'},
-  {label: 'H4', style: 'header-four', icon:'fa fa-header'},
-  {label: 'H5', style: 'header-five', icon:'fa fa-header'},
-  {label: 'H6', style: 'header-six', icon:'fa fa-header'},
+  {label: 'H1', style: 'header-one', icon:''},
+  {label: 'H2', style: 'header-two', icon:''},
+  {label: 'H3', style: 'header-three', icon:''},
+  {label: 'H4', style: 'header-four', icon:''},
+  {label: 'H5', style: 'header-five', icon:''},
+  {label: 'H6', style: 'header-six', icon:''},
   {label: 'Blockquote', style: 'blockquote', icon:'fa fa-quote-right'},
   {label: 'UL', style: 'unordered-list-item', icon:'fa fa-list-ul'},
   {label: 'OL', style: 'ordered-list-item', icon:'fa fa-list-ol'},
