+++ src/components/editor/Editor.js
@@ -4,6 +4,7 @@ import {Editor,
         RichUtils, 
         DraftEditorBlock,
         DefaultDraftBlockRenderMap,
+        convertToRaw,
       } from 'draft-js';
 import InlineStyleControls from './InlineStyleControls';
 import BlockStyleControls from './BlockStyleControls';
@@ -67,6 +68,10 @@ class MyEditor extends Component {
     this.onTab = (e) => this._onTab(e);
     this.toggleBlockType = (type) => this._toggleBlockType(type);
     this.toggleInlineStyle = (style) => this._toggleInlineStyle(style);
+    this.logState = () => {
+      const content = this.state.editorState.getCurrentContent();
+      console.log(convertToRaw(content));
+    };
   }
   
   _onBoldClick() {
@@ -133,7 +138,7 @@ class MyEditor extends Component {
               <button onClick={this._onItalicClick.bind(this)}>Italic</button>
               <button onClick={this._onStrikThroughClick.bind(this)}>StrikThrough</button>
               <button onClick={this._onCodeClick.bind(this)}>Code</button>
-              
+              <button onClick={this.logState}>Content</button>
               <BlockStyleControls
                 editorState={editorState}
                 onToggle={this.toggleBlockType}
